utilities/requirements.txt
autogen==0.3.1
beautifulsoup4==4.12.3
certifi==2024.7.4
docx2txt==0.8
groq==0.11.0
llama_index==0.11.17
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-groq==0.2.0
llama-index-retrievers-bm25==0.4.0
llama-index-readers-web==0.2.4
llama_parse==0.5.7
pdf2docx==0.5.8
pdfplumber==0.11.2
pdfplumber==0.11.3
python-dotenv==1.0.1
python_docx==1.1.2
python_pptx==1.0.2
Requests==2.32.3
streamlit==1.37.0
streamlit==1.37.1
